feat: перенос проекта на tsx, подключение пакета
