+++ src/App.tsx
@@ -1,0 +1,11 @@
+import { DocumentVisibility } from "./UseDocumentVisibility/UseDocumentVisibility";
+
+function App() {
+  return (
+    <div className="App">
+      <DocumentVisibility />
+    </div>
+  );
+}
+
+export default App;
